add close method to client
